Implement fractional_scale_v1 and wp_viewport
This allows compositors to hint a non-integer scale to use on a window
which we can hook to Qt's existing fractional scaling support.

The viewport is used to communicate the relationship between buffer size
and logical size to the compositor. It is a non-integer alternative to
wl_buffer_scale

Change-Id: I1a850f1bcd40e8d04e241e18a538b11f18bc671c
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: David Edmundson <davidedmundson@kde.org>

Backport to Qt5, disable check for highDpiScaleFactorRoundingPolicy and remove tests:
Signed-off-by: Dobroslaw Kijowski <dobo90@gmail.com>
(cherry picked from commit 23d3fc712cdf7fc0a4ff837082de9ba773e8605c)